dev-requirements.txt
# documentation tools
sphinx>=1.3
sphinxcontrib-seqdiag
Flask-Sphinx-Themes
# testing tools
pytest
pytest-cov
mock
responses
freezegun
# testing sqlalchemy support
sqlalchemy>=0.9
sqlalchemy_utils
flask-sqlalchemy
# testing integration with other extensions
flask-login
flask-caching
betamax
